+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 Changelog
 =========
-## TBA
-
+## 0.1.0
+- Fix #4: Throwing an error when the last edited file was deleted from outside the module
+- Enh: Warning if the requested module was not found